ci: security scanners, reproducible release, canonical signing encoder
Adds the CI pipeline and the security work it surfaced.

Pipeline (scripts/dev/ci.sh, mirrored in .github/workflows/ci.yml):
gofmt, go vet, staticcheck, gosec, gitleaks, govulncheck, the logging and
dashboard policy checks, and go test -race. The same script runs locally and
in Actions so the two cannot drift.

Findings fixed rather than suppressed:

- Four modules each had their own copy of the signing preimage framing
  logic. Consolidated into app/canonical, which enforces the length bound
  that all four previously assumed. Framing drift in signing code is
  signature confusion, so one implementation with tests is the point.
- The outer domain-separation wrapper and Merkle challenge responses frame
  arbitrary-length payloads, so they now use 64-bit length prefixes, which
  cannot overflow, instead of 32-bit prefixes that needed a check with no
  meaningful handling at that layer.
- Replaced a hand-rolled Sscanf hex parser in hashgram-keygen with
  encoding/hex.
- Split SelectIndex out of ChallengeChunkIndex: the two answer different
  questions and reusing one for both meant two unbounded narrowings in
  consensus code.
- Guarded the statfs block size, where a filesystem reporting nonsense would
  have produced an enormous free-space figure in preflight output.
- staticcheck caught a dead check in that same guard: Statfs_t.Blocks and
  Bavail are already uint64 on Linux.
- Upgraded ten dependencies to close thirteen symbol-level advisories in the
  shipped binary, down to zero unreviewed.
- Verified that GO-2024-2584 is a stale advisory entry: the slashing evasion
  it describes was patched in SDK 0.50.5 and this tree is on v0.53.8.

Reproducible release (scripts/dev/release.sh): CGO off, -trimpath,
-buildvcs=false, no build-date or commit stamp. Verified byte-identical
across builds with the cache cleared. CGO_ENABLED=0 was tested by running a
devnet from the resulting binary, which contradicted the Makefile comment
claiming CGO was required; the comment is corrected and the binaries are 27
percent smaller and static.

Observability: 22 hashgram_* Prometheus gauges emitted from EndBlock covering
supply against its ceiling, the Founder ledger, the service reserve and the
welcome pool; three Grafana dashboards; 24 alert rules. Metric names were
taken from a running node rather than from memory, which is how the missing
peers gauge and the absent missed-blocks metric were found.

Also adds tools/tokenomics-simulator, which imports the chain's own emission,
tier and FounderCut functions so its curves are the ones the state machine
runs, and docs/LOGGING_POLICY.md with a checker that self-tests its patterns.

+++ cmd/hashgramctl/cmd/backup.go
@@ -118,7 +118,7 @@ live database produces a corrupt copy.`,
 
 			o := newOut()
 			o.row("Archive", archive)
-			o.row("Size", hgsys.HumanBytes(uint64(info.Size())))
+			o.row("Size", hgsys.HumanBytes(info.Size()))
 			o.row("Chain data included", includeChainDB)
 			o.blank()
 			o.raw("EXCLUDED from this archive, on purpose:")
